CoinGeckoAPI

Documentation for CoinGeckoAPI.

CoinGeckoAPI._api_url_params

CoinGeckoAPI.get_asset_platforms

CoinGeckoAPI.get_coin_by_id

CoinGeckoAPI.get_coin_history_by_id

CoinGeckoAPI.get_coin_info_from_contract_address_by_id

CoinGeckoAPI.get_coin_market_chart_by_id

CoinGeckoAPI.get_coin_market_chart_from_contract_address_by_id

CoinGeckoAPI.get_coin_market_chart_range_by_id

CoinGeckoAPI.get_coin_market_chart_range_from_contract_address_by_id

CoinGeckoAPI.get_coin_ohlc_by_id

CoinGeckoAPI.get_coin_status_updates_by_id

CoinGeckoAPI.get_coin_ticker_by_id

CoinGeckoAPI.get_coins_categories

CoinGeckoAPI.get_coins_categories_list

CoinGeckoAPI.get_coins_list

CoinGeckoAPI.get_coins_markets

CoinGeckoAPI.get_exchanges_list

CoinGeckoAPI.get_price

CoinGeckoAPI.get_supported_vs_currencies

CoinGeckoAPI.ping

CoinGeckoAPI._api_url_params — Function

_api_url_params(api_url, params, api_url_has_params = false)

Function to build the API url with the parameters

Arguments

`api_url: string` : the API url
`params: dict` : the parameters to be added to the API url
`api_url_has_params: bool` : if the API url has parameters

Returns

`api_url: String` : the API url with the parameters

CoinGeckoAPI.get_asset_platforms — Method

get_asset_platforms()

List all the asset platforms (blockchain networks)

Arguments

None

Returns

`asset_platforms: Dict` : the asset platforms

CoinGeckoAPI.get_coin_by_id — Method

get_coin_by_id(id, kargs...)

Get current data (name, price, market, including exchange tickers for a coin)

Arguments

`id: string` : the coin id 
`kargs: dict` : the parameters to be added to the API url

Returns

`coin: Dict` : the current data for a coin

CoinGeckoAPI.get_coin_history_by_id — Method

get_coin_history_by_id(id, date, kargs...)

Get Historical data (name, price, market, stats ) at a given date for a coin.

Arguments

`id: string` : the coin id 
`date: string` : the date to get the data for
`kargs: dict` : the parameters to be added to the API url

Returns

`coin_history: Dict` : the historical data for a coin

Example:

```julia r = getcoinhistorybyid("bitcoin", "30-12-2017")

````

CoinGeckoAPI.get_coin_info_from_contract_address_by_id — Method

get_coin_info_from_contract_address_by_id(id, contract_address)

Get coin info from contract address

Arguments

`id: string` : Asset platform (See asset_platforms endpoint for list of options)
`contract_address: string` : the token's contract address

Returns

`coin_info: Dict` : the coin info from contract address

Example:

get_coin_info_from_contract_address_by_id("ethereum", "0xc00e94cb662c3520282e6f5717214004a7f26888")

CoinGeckoAPI.get_coin_market_chart_by_id — Method

get_coin_market_chart_by_id(id, vs_currency, days, kargs...)

Get historical market data include price, market cap, and 24h volume (granularity auto) Minutely data will be used for duration within 1 day, Hourly data will be used for duration between 1 day and 90 days, Daily data will be used for duration above 90 days.

Arguments

`id: string` : the coin id 
`vs_currency: string` : the currency to get the data in
`days: String` : the number of days to get the data for
`kargs: dict` : the parameters to be added to the API url

Returns

`coin_market_chart: Dict` : the historical market data for a coin

Example:

r = get_coin_market_chart_by_id("bitcoin", "usd", "30")

CoinGeckoAPI.get_coin_market_chart_from_contract_address_by_id — Method

get_coin_market_chart_from_contract_address_by_id(id, contract_address, vs_currency, days)

Get historical market data include price, market cap and 24h volume (granularity auto)

Arguments

`id: string` : The id of the platform issuing tokens (See asset_platforms endpoint for list of options)
`contract_address: string` : Token's contract address
`vs_currency: string` : the currency to get the data in
`days: String` : Data up to number of days ago (eg. 1,14,30,max)

Returns

`coin_market_chart: Dict` : the historical market data for a coin

Example:

r = get_coin_market_chart_from_contract_address_by_id("ethereum", "0xc00e94cb662c3520282e6f5717214004a7f26888", "usd", "1")

CoinGeckoAPI.get_coin_market_chart_range_by_id — Method

get_coin_market_chart_range_by_id(id, vs_currency, from_timestamp, to_timestamp, kargs...)

Get historical market data include price, market cap, and 24h volume within a range of timestamp (granularity auto) Data granularity is automatic (cannot be adjusted) 1 day from query time = 5 minute interval data 1 - 90 days from query time = hourly data above 90 days from query time = daily data (00:00 UTC)

Arguments

`id: string` : the coin id 
`vs_currency: string` : the currency to get the data in
`from_timestamp: String` : the timestamp to start the data from (UNIX Timestamp (eg 1392577232))
`to_timestamp: String` : the timestamp to end the data from (UNIX Timestamp (eg 1422577232))
`kargs: dict` : the parameters to be added to the API url

Returns

`coin_market_chart: Dict` : the historical market data for a coin

Example:

r = get_coin_market_chart_range_by_id("bitcoin", "usd", "1392577232", "1422577232")

CoinGeckoAPI.get_coin_market_chart_range_from_contract_address_by_id — Method

get_coin_market_chart_range_from_contract_address_by_id(id, contract_address, vs_currency, from_timestamp, to_timestamp)

Get historical market data include price, market cap, and 24h volume within a range of timestamp (granularity auto)

Arguments

`id: string` : The id of the platform issuing tokens (See asset_platforms endpoint for list of options)
`contract_address: string` : Token's contract address
`vs_currency: string` : the currency to get the data in
`from_timestamp: String` : From date in UNIX Timestamp (eg. 1392577232)
`to_timestamp: String` : To date in UNIX Timestamp (eg. 1422577232)

Returns

`coin_market_chart: Dict` : the historical market data for a coin

CoinGeckoAPI.get_coin_ohlc_by_id — Method

get_coin_ohlc_by_id(id, vs_currency, days)

Get coin's ohlc Candles's body: 1-2 days: 30 minutes 3-30 days: 4 hours 31 and before: 4 days

Arguments

`id: string` : the coin id  (eg bitcoin)
`vs_currency: string` : the currency to get the data in
`days: String` : the number of days to get the data for
`kargs: dict` : the parameters to be added to the API url

Returns

`coin_ohlc: Dict` : the ohlc data for a coin

Example:

r = get_coin_ohlc_by_id("bitcoin", "usd", "30")

CoinGeckoAPI.get_coin_status_updates_by_id — Method

get_coin_status_updates_by_id(id, kargs...)

Get status updates for a given coin.

Arguments

`id: string` : the coin id  (eg bitcoin)
`kargs: dict` : the parameters to be added to the API url

Returns

`coin_status_updates: Dict` : the status updates for a coin

CoinGeckoAPI.get_coin_ticker_by_id — Method

get_coin_ticker_by_id(id, kargs...)

Get coin tickers (paginated to 100 items)

Arguments

`id: string` : the coin id 
`kargs: dict` : the parameters to be added to the API url

Returns

`coin_ticker: Dict` : the coin tickers

CoinGeckoAPI.get_coins_categories — Method

get_coins_categories(kargs...)

List all categories with market data.

Arguments

`kargs: Dict` : the parameters to be added to the API url
valid values: market_cap_desc (default), market_cap_asc, 
name_desc, name_asc, market_cap_change_24h_desc and market_cap_change_24h_asc

Returns

`coins_categories: Dict` : the coins categories

Example:

```julia getcoinscategories() getcoinscategories("marketcapdesc" => "true")

CoinGeckoAPI.get_coins_categories_list — Method

get_coins_categories_list()

List all the coins categories

Arguments

None

Returns

`coins_categories: Dict` : the coins categories

CoinGeckoAPI.get_coins_list — Method

get_coins(kargs...)

List all coins with data (name, price, market, developer, community, etc)

Arguments

`kargs: dict` : the parameters to be added to the API url

Returns

`coins: Dict` : the list of coins with id, name and symbol

CoinGeckoAPI.get_coins_markets — Method

get_coins_markets(vs_currencies, kargs...)

Use this to obtain all the coins market data (price, market cap, volume)

Arguments

`vs_currencies: list` : the currencies to get the price in
`kargs: dict` : the parameters to be added to the API url

Returns

`coins_markets: Dict` : the list of coins with market data

Example:

```julia r = getcoinsmarkets("usd")

````

CoinGeckoAPI.get_exchanges_list — Method

function get_exchanges_list(kargs...)

List all exchanges.

Arguments

`kargs: Dict` : the parameters to be added to the API url
valid values: 
per_page : number of results per page 1..250. Total results per page. Default value is 100
page : Page through results.

Returns

`exchanges_list: Dict` : the exchanges list

Example:

get_exchanges_list("per_page" => "10")

CoinGeckoAPI.get_price — Method

get_price(ids, vs_currencies, kargs...)

Get the price of a cryptocurrency in any other supported currencies needed

Arguments

`ids: list` : the cryptocurrency ids
`vs_currencies: list` : the currencies to get the price in
`kargs: dict` : the parameters to be added to the API url

Returns

`prices: dict` : the prices of the cryptocurrencies in the currencies

Example:

r = get_price("bitcoin", "usd", "include_market_cap" => "true", "include_24hr_vol" => "true", "include_24hr_change" => "true")

CoinGeckoAPI.get_supported_vs_currencies — Method

get_supported_vs_currencies()

Get list of supportedvscurrencies

Arguments

None

Returns

`supported_vs_currencies: Dict` : the list of supported_vs_currencies

CoinGeckoAPI.ping — Method

ping()

Function to check if the API is up and running

Arguments

None

Returns